requirements.txt
sentence-transformers>=2.2.2
spacy>=3.7.2
hdbscan>=0.8.33
numpy>=1.26.2
pandas>=2.1.3
scipy>=1.11.4
pymannkendall>=1.4.3
supabase>=2.3.4
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.1.0
# --- Microservice ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
openai>=1.0.0
scikit-learn>=1.3.0
